pyproject.toml
[tool.poetry]
name = "discordbattlebot"
version = "0.1.0"
description = ""
authors = ["Lifelite <j.t.huntsman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
mysqlclient = "^2.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
